refactor(content): update datetime structure
