+++ src/lib/components/datetime/index.ts
@@ -1,0 +1,1 @@
+export { default as Datetime } from "./datetime.svelte";
