+++ src/lib/components/index.ts
@@ -2,7 +2,7 @@ export { Button } from "./button";
 export { Canvas, CanvasLayer } from "./canvas";
 export { CardArticle, CardProject, CardSketch } from "./card";
 export { default as Code } from "./code/Code.svelte";
-export { default as Datetime } from "./Datetime.svelte";
+export { Datetime } from "./datetime";
 export { Details } from "./details";
 export { Dialog } from "./dialog";
 export { Drawer } from "./drawer";
